Bumped to 0.3.0
- Fix compatibility against the latest NCBI assembly report file.
- Revise the filtering rule for NCBI ANI report file.
- Add dump_sqlite_db command to dqc_admin_tool.py to dump the list of reference genomes to file.
- Add status to taxonomy check result
	- __conclusive__: Effective ANI hit (>=95%) againt only 1 species, hence the species name is conclusively determined.
	- __indistinguishable__: The genome belongs to one of the species that are difficult to distinguish using ANI (e.g. E. coli and Shigella spp.)
	- __inconsistent__: ANI hits against more than 2 differenct species. This may result from the comparison between very closely-related species or contamination of 2 different species.
	- __below_threhold__: The ANI hit is below the threshold (95%)

+++ dqc/admin/download_all_reference_genomes.py
@@ -11,9 +11,11 @@ logger = get_logger(__name__)
 def delete_unwanted_genomes(accessions, genome_dir):
     for accession in accessions:
         target_file = os.path.join(genome_dir, accession + ".fna.gz")
-        assert os.path.exists(target_file)
-        os.remove(target_file)
-        logger.info("Deleted %s", target_file)
+        if not os.path.exists(target_file):
+            logger.warning("Target file does not exist. [%s]", target_file)
+        else:
+            os.remove(target_file)
+            logger.info("Deleted %s", target_file)
 
 def get_existing_genomes(genome_dir):
     # Todo: check eof, broken file
